requirements.txt
APScheduler==3.6.3
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.2
boto3==1.16.63
botocore==1.19.63
bs4==0.0.1
bson==0.5.10
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.58
confuse==1.4.0
constantly==15.1.0
countryinfo==0.1.2
cryptography==3.4.6
cssselect==1.1.0
decorator==5.1.0
dnspython==1.16.0
elementpath==2.2.3
etsy2==0.7.0
firebase-admin==4.3.0
flasgger==0.9.4
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.8
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.5.2
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.0
mysql-connector==2.2.9
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
parsel==1.6.0
phpserialize==1.3
Pillow==7.2.0
Protego==0.1.16
protobuf==3.12.2
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.9.8
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pykerberos==1.2.1
pymongo==3.11.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-amazon-sp-api==0.6.2
python-dateutil==2.8.1
python-http-client==3.3.3
python-snappy==0.5.4
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.7
Scrapy==2.4.1
sendgrid==6.9.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
starkbank-ecdsa==1.1.1
Twisted==20.3.0
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
validators==0.18.2
w3lib==1.22.0
Werkzeug==1.0.1
WooCommerce==3.0.0
xmljson==0.2.1
xmlschema==1.6.4
xmltodict==0.12.0
zipp==3.4.0
zope.interface==5.2.0
